# working-scripts/outputs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MVE Resource Server API",
    "version": "1.0.0",
    "description": "MVE (Meta Virtual Environment) Resource Server API\n\n이 API는 다음 기능을 제공합니다:\n- 🎵 음원 파일 관리 (스트리밍, 업로드, 검색)\n- 🎨 3D 모델 파일 관리 (GLB)\n- 🤖 AI 기반 3D 모델 생성\n- 🎭 아바타 악세서리 프리셋 관리\n- 🎪 콘서트 세션 관리\n\n**인증**: 모든 API는 JWT Bearer 토큰 인증이 필요합니다.\n**토큰 발급**: [MVE Login Server](https://github.com/prayslaks/mve-login-server)에서 로그인하여 JWT 토큰을 발급받으세요.",
    "contact": {
      "name": "MVE Development Team",
      "url": "https://github.com/prayslaks"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "로컬 개발 서버"
    },
    {
      "url": "http://your-ec2-public-ip",
      "description": "AWS EC2 프로덕션 서버 (HTTP)"
    },
    {
      "url": "https://your-domain.com",
      "description": "AWS EC2 프로덕션 서버 (HTTPS)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "MVE Login Server에서 발급받은 JWT 토큰을 입력하세요."
      }
    },
    "schemas": {
      "Vector3D": {
        "type": "object",
        "description": "3D 벡터 (위치)",
        "required": [
          "x",
          "y",
          "z"
        ],
        "properties": {
          "x": {
            "type": "number",
            "description": "X 좌표",
            "example": 0
          },
          "y": {
            "type": "number",
            "description": "Y 좌표",
            "example": 0
          },
          "z": {
            "type": "number",
            "description": "Z 좌표",
            "example": 0
          }
        }
      },
      "Rotator": {
        "type": "object",
        "description": "3D 회전 (Pitch, Yaw, Roll)",
        "required": [
          "pitch",
          "yaw",
          "roll"
        ],
        "properties": {
          "pitch": {
            "type": "number",
            "description": "Pitch (상하 회전)",
            "example": 0
          },
          "yaw": {
            "type": "number",
            "description": "Yaw (좌우 회전)",
            "example": 0
          },
          "roll": {
            "type": "number",
            "description": "Roll (롤 회전)",
            "example": 0
          }
        }
      },
...
      "Accessory": {
        "type": "object",
        "description": "아바타 액세서리",
        "required": [
          "socketName",
          "relativeLocation",
          "relativeRotation",
          "modelUrl"
        ],
        "properties": {
          "socketName": {
            "type": "string",
            "description": "소켓 이름",
            "example": "hand_socket"
          },
          "relativeLocation": {
            "$ref": "#/components/schemas/Vector3D"
          },
          "relativeRotation": {
            "$ref": "#/components/schemas/Rotator"
          },
          "relativeScale": {
            "type": "number",
            "format": "float",
            "description": "상대적 스케일",
            "example": 1
          },
          "modelUrl": {
            "type": "string",
            "description": "3D 모델 URL",
            "example": "https://example.com/models/microphone.glb"
          }
        }
      },
      "AccessoryPreset": {
        "type": "object",
        "description": "액세서리 프리셋",
        "properties": {
          "id": {
            "type": "integer",
            "description": "프리셋 ID",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "description": "소유자 사용자 ID",
            "example": 1
          },
          "presetName": {
            "type": "string",
            "description": "프리셋 이름",
            "example": "My Preset"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "프리셋 설명",
            "example": "My favorite accessories"
          },
          "accessories": {
            "type": "array",
            "description": "액세서리 목록",
            "items": {
              "$ref": "#/components/schemas/Accessory"
            }
          },
          "isPublic": {
            "type": "boolean",
            "description": "공개 여부",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "생성 시간"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "수정 시간"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/presets/save": {
      "post": {
        "summary": "액세서리 프리셋 저장",
        "description": "아바타 액세서리 프리셋을 저장합니다",
        "tags": [
          "Accessory Presets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "presetName",
                  "accessories"
                ],
                "properties": {
                  "presetName": {
                    "type": "string",
                    "example": "My Preset"
                  },
                  "accessories": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Accessory"
                    }
                  },
                  "description": {
                    "type": "string"
                  },
                  "isPublic": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프리셋 저장 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessoryPreset"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          }
        }
      }
    },
    "/api/presets/list": {
      "get": {
        "summary": "프리셋 목록 조회",
        "description": "사용자의 액세서리 프리셋 목록을 조회합니다",
        "tags": [
          "Accessory Presets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "includePublic",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "공개 프리셋 포함 여부"
          }
        ],
        "responses": {
          "200": {
            "description": "프리셋 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "presets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AccessoryPreset"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/presets/{id}": {
      "get": {
        "summary": "프리셋 상세 조회",
        "description": "특정 프리셋의 상세 정보를 조회합니다 (액세서리 데이터 포함)",
        "tags": [
          "Accessory Presets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "프리셋 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "프리셋 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "preset": {
                      "$ref": "#/components/schemas/AccessoryPreset"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "접근 권한 없음"
          },
          "404": {
            "description": "프리셋을 찾을 수 없음"
          }
        }
      },
      "put": {
        "summary": "프리셋 업데이트",
        "description": "프리셋 정보를 업데이트합니다",
        "tags": [
          "Accessory Presets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "프리셋 ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "presetName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "accessories": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Accessory"
                    }
                  },
                  "isPublic": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프리셋 업데이트 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessoryPreset"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "접근 권한 없음"
          },
          "404": {
            "description": "프리셋을 찾을 수 없음"
          },
          "409": {
            "description": "중복된 프리셋 이름"
          }
        }
      },
      "delete": {
        "summary": "프리셋 삭제",
        "description": "프리셋을 삭제합니다",
        "tags": [
          "Accessory Presets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "프리셋 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "프리셋 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "접근 권한 없음"
          },
          "404": {
            "description": "프리셋을 찾을 수 없음"
          }
        }
      }
    },
...
  }
}